requirements.txt
azure.iot.device 